requirements.txt
streamlit
pandas
numpy
matplotlib
ufal.udpipe
compress-fasttext
networkx